+++ src/MessageTranslator.php
@@ -3,14 +3,12 @@
  * UserFrosting (http://www.userfrosting.com)
  *
  * @link      https://github.com/userfrosting/i18n
- * @copyright Copyright (c) 2013-2017 Alexander Weissman
- * @license   https://github.com/userfrosting/i18n/blob/master/licenses/UserFrosting.md (MIT License)
+ * @license   https://github.com/userfrosting/UserFrosting/blob/master/LICENSE.md (MIT License)
  */
-
 namespace UserFrosting\I18n;
 
-use Illuminate\Config\Repository;
 use UserFrosting\Support\Exception\FileNotFoundException;
+use UserFrosting\Support\Repository\Repository;
 
 /**
  * MessageTranslator Class
@@ -18,163 +16,31 @@ use UserFrosting\Support\Exception\FileNotFoundException;
  * Translate message ids to a message in a specified language.
  *
  * @author    Louis Charette
- * @author    Alexander Weissman
+ * @author    Alexander Weissman (https://alexanderweissman.com)
  */
 class MessageTranslator extends Repository
 {
     /**
-     * @var array an array of paths to search for locale files.
-     */
-    protected $paths = [];
-
-    /**
-     * @var Twig_Environment A Twig envrionnement used to replace placeholders.
+     * @var Twig_Environment A Twig environment used to replace placeholders.
      */
     protected $twig;
 
     /**
-     * @var array an array of paths to search for locale files.
+     * @var string The default key that contains the pluralization code.
      */
     protected $defaultPluralKey = 'plural';
 
+    /**
+     * Create the translator.
+     *
+     * @param array $items
+     */
     public function __construct(array $items = [])
     {
         $this->items = $items;
 
         $loader = new \Twig_Loader_Filesystem();
         $this->twig = new \Twig_Environment($loader);
-    }
-
-    /**
-     * Add a path to search for locale files.
-     *
-     * @param string $path
-     */
-    public function addPath($path)
-    {
-        if (!is_dir($path)) {
-            throw new FileNotFoundException("The locale path '$path' could not be found or is not readable.");
-        }
-
-        $this->paths[] = $path;
-    }
-
-    /**
-     * Set an array of paths to search for locale files.
-     *
-     * @param array $paths
-     */
-    public function setPaths(array $paths = [])
-    {
-        $this->paths = $paths;
-    }
-
-    /**
-     * Return a list of all paths to search for locale files containing the translation table to be used.  A translation table is an associative array of message ids => translated messages.
-     *
-     * @return array
-     */
-    public function getPaths()
-    {
-        return $this->paths;
-    }
-
-    /**
-     * Recursively merge locale values (scalar or array) into this repository.
-     *
-     * If no key is specified, the items will be merged in starting from the top level of the array.
-     * If a key IS specified, items will be merged into that key.
-     * Nested keys may be specified using dot syntax.
-     * @param string|null $key
-     * @param mixed $items
-     */
-    public function mergeItems($key = null, $items)
-    {
-        $targetValues = array_get($this->items, $key);
-
-        if (is_array($targetValues)) {
-            $modifiedValues = array_replace_recursive($targetValues, $items);
-        } else {
-            $modifiedValues = $items;
-        }
-
-        array_set($this->items, $key, $modifiedValues);
-    }
-
-    /**
-     * Recursively merge a locale file into this repository.
-     *
-     * @param string $fileWithPath
-     */
-    public function mergeLocaleFile($fileWithPath)
-    {
-        if (!(file_exists($fileWithPath) && is_readable($fileWithPath))) {
-            throw new FileNotFoundException("The locale file '$fileWithPath' could not be found or is not readable.");
-        } else {
-            // Use null key to merge the entire locale array
-            $this->mergeItems(null, require $fileWithPath);
-        }
-    }
-
-    /**
-     * Load the locale items from all of the files.
-     *
-     * @param string|array|null $localeArray
-     */
-    public function loadLocaleFiles($localeArray = [])
-    {
-        //So we can accept strings argument also
-        if (!is_array($localeArray)) {
-            $localeArray = array($localeArray);
-        }
-
-        foreach ($localeArray as $locale) {
-            // Search each locale path for default and environment-specific locale files
-            foreach ($this->paths as $path) {
-                // Merge in default locale file
-                $defaultFiles = $this->getLocaleFiles($path, $locale);
-
-                foreach ($defaultFiles as $fileWithPath) {
-                    $this->mergeLocaleFile($fileWithPath);
-                }
-            }
-        }
-    }
-
-    /**
-     * Get an array of full paths found in a locale directory
-     *
-     * @param string $path
-     * @param string $locale
-     *
-     * @return array List of file found at this path
-     */
-    protected function getLocaleFiles($path, $locale)
-    {
-        // Find all the php files in the locale directory
-        $filesWithPath = glob(rtrim($path, '/\\') . '/' . $locale . "/*.php");
-
-        return $filesWithPath;
-    }
-
-    /**
-     * Get an array of available locales in currently defined paths
-     *
-     * @param none
-     *
-     * @return array List of locales found in current paths
-     */
-    public function getAvailableLocales()
-    {
-        $availableLocales = [];
-
-        foreach ($this->paths as $path) {
-            $fullPaths = glob(rtrim($path, '/\\') . '/*', GLOB_ONLYDIR);
-            $locales = array_map('basename', $fullPaths);
-            $availableLocales = array_replace_recursive($availableLocales, $locales);
-        }
-
-        return $availableLocales;
     }
 
     /**